Create bubble_short.cpp
This program defines a bubbleSort function that sorts an array of integers in ascending order using the bubble sort algorithm. In the main function, it takes input from the user for the number of elements and the array elements, sorts the array using bubbleSort, and then displays the sorted array.